// src/utils/bagisto/types.ts
export type selectedPaymentMethodType = {
  id: string;
  method: string;
  methodTitle: string;
  cartId: string;
  methodDescription?:string
};

export type ChannelType = {
  id: string;
  code: string;
  name: string;
  description: string;
  theme: string;
  hostname: string;
  defaultLocaleId: number;
  baseCurrencyId: number;
  rootCategoryId: number;
  logoUrl: string;
  faviconUrl: string;
  success: string;
};

export type BagistoChannelOperation = {
  data: {
    getDefaultChannel: ChannelType;
  };
};

export type CartItem = {
  id: string;
  type: string;
  quantity: number;
  sku: string;
  name: string;
  couponCode: string;
  weight: string;
  totalWeight: string;
  baseTotalWeight: string;
  price: string;
  basePrice: string;
  total: string;
  baseTotal: string;
  taxPercent: string;
  taxAmount: string;
  baseTaxAmount: string;
  discountPercent: string;
  discountAmount: string;
  baseDiscountAmount: string;
  parentId: string;
  productId: string;
  cartId: string;
  taxCategoryId: string;
  customPrice: string;
  appliedCartRuleIds: string;
  createdAt: string;
  updatedAt: string;
  product: {
    id: string;
    type: string;
    name: string;
    urlKey: string;
    attributeFamilyId: string;
    shortDescription: string;
    guestCheckout: boolean;
    sku: string;
    parentId: string;
    variants: {
      id: string;
      type: string;
      attributeFamilyId: string;
      sku: string;
      parentId: string;
    };
    parent: {
      id: string;
      type: string;
      attributeFamilyId: string;
      sku: string;
      parentId: string;
    };
    cacheBaseImage: {
      smallImageUrl: string;
      mediumImageUrl: string;
      largeImageUrl: string;
      originalImageUrl: string;
    }[];
    attributeValues: {
      id: string;
      productId: string;
      attributeId: string;
      locale: string;
      channel: string;
      textValue: string;
      booleanValue: string;
      integerValue: string;
      floatValue: string;
      dateTimeValue: string;
      dateValue: string;
      jsonValue: string;
      attribute: {
        id: string;
        code: string;
        adminName: string;
        type: string;
      };
    };
    superAttributes: {
      id: string;
      code: string;
      adminName: string;
      type: string;
      position: string;
    };
    inventories: {
      id: string;
      qty: string;
      productId: string;
      inventorySourceId: string;
      vendorId: string;
    };
    images: {
      id: string;
      url: string;
      type: string;
      path: string;
      productId: string;
    }[];
  };
  formattedPrice: {
    price: string;
    basePrice: string;
    total: string;
    baseTotal: string;
    taxAmount: string;
    baseTaxAmount: string;
    discountAmount: string;
    baseDiscountAmount: string;
  };
  payment: selectedPaymentMethodType;
};
// ...
export type BagistoCart = {
  id: string;
  type: string;
  customerEmail: string;
  customerFirstName: string;
  customerLastName: string;
  shippingMethod: string;
  couponCode: string;
  itemsCount: string;
  itemsQty: string;
  cartCurrencyCode: string;
  grandTotal: string;
  baseGrandTotal: string;
  subTotal: string;
  baseSubTotal: string;
  taxTotal: string;
  baseTaxTotal: string;
  discountAmount: string;
  baseDiscountAmount: string;
  checkoutMethod: string;
  isGuest: boolean;
  isActive: string;
  items: Array<CartItem>;
  product: BagistoProductInfo;
  payment?: selectedPaymentMethodType;
  selectedShippingRate: {
    price: string;
    method: string;
  };
  shippingAddress: AddressDataTypes;
  billingAddress: AddressDataTypes;
};
// ...
export type BagistoProductInfo = {
  name: string;
  longDescription?: string;
  urlKey: string;
  type: string;
  status: boolean;
  slug?: string;
  width: string;
  height: string;
  path?: string;
  metaTitle: string;
  metaDescription: string;
  price: ProductPrice;
  id: string;
  handle: string;
  availableForSale: boolean;
  title: string;
  description: string;
  shortDescription: string;
  descriptionHtml: string;
  options: ProductOption[];
  priceRange: {
    maxVariantPrice: Money;
    minVariantPrice: Money;
  };
  cacheGalleryImages: {
    smallImageUrl: string;
    mediumImageUrl: string;
    largeImageUrl: string;
    originalImageUrl: string;
  }[];
  priceHtml: {
    regularPrice: string;
    currencyCode: string;
    formattedRegularPrice: string;
    finalPrice: string;
    formattedFinalPrice: string;
  };
  variants: Array<ProductVariant>;
  featuredImage: Image;
  images: Array<ImageInfo>;
  productFlats: {
    id: string;
    name: string;
    description: string;
    metaTitle: string;
    metaDescription: string;
    width: string;
    height: string;
  }[];
  relatedProducts: RelatedProducts[];
  inventories: {
    qty: string;
  }[];
  configutableData: {
    attributes: ConfigurableProductData[];
    index: ConfigurableProductIndexData[];
  };
  seo: SEO;
  tags: string[];
  updatedAt: string;
};
// ...
export type AddressDataTypes = {
  id: string;
  firstName: string;
  lastName: string;
  companyName: string;
  address: string;
  city: string;
  state: string;
  countryName: string;
  postcode: string;
  email: string;
  country: string;
  addressType: string;
  vatId: string;
  phone: string;
  stateName: string;
  defaultAddress: boolean;
  useForShipping: boolean;
};
// ...
export type BagistoUpdateCartOperation = {
  data: {
    updateItemToCart: {
      cart: BagistoCart;
    };
  };
  variables: {
    input: {
      qty: {
        cartItemId: number;
        quantity: number;
      }[];
    };
  };
};